requirements.txt
asgiref==3.2.10
Django==3.1.1
django-grappelli==2.14.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-mptt-admin==2.0.0
mysql-connector==2.2.9
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
